requirements.txt
Babel==2.9.0
cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-API==1.1
Flask-Assets==0.12
Flask-Babel==0.11.2
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Session==0.3.2
Flask-WTF==0.14.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.5
requests==2.24.0
six==1.15.0
urllib3==1.25.9
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.3
Pillow==9.0.1
